test(suite): relocate desktop dependencies mock
