+++ packages/suite/src/components/suite/SecurityCheck/DeviceCompromised.test.tsx
@@ -8,10 +8,10 @@ import { DeviceModelInternal } from '@trezor/device-utils';
 
 import { type AppState } from 'src/reducers/store';
 import { configureStore } from 'src/support/tests/configureStore';
-import { extraDependenciesDesktopMock } from 'src/support/tests/extraDependenciesDesktop.mock';
 import { renderWithProviders } from 'src/support/tests/hooksHelper';
 
 import { DeviceCompromised } from './DeviceCompromised';
+import { extraDependenciesDesktopMock } from '../../../../mocks/extraDependenciesDesktopMock';
 import { mockInitialAppState } from '../../../../mocks/mockInitialAppState';
 
 jest.mock('@suite-common/tx-simulation', () => ({}));
